+++ projects/core/src/remult3/RepositoryImplementation.ts
@@ -98,11 +98,11 @@ import { addValidator } from './addValidator.js'
 
 let classValidatorValidate:
   | ((
-    item: any,
-    ref: {
-      fields: FieldsRef<any>
-    },
-  ) => Promise<void>)
+      item: any,
+      ref: {
+        fields: FieldsRef<any>
+      },
+    ) => Promise<void>)
   | undefined = undefined
 // import ("class-validator".toString())
 //     .then((v) => {
@@ -221,7 +221,7 @@ export class RepositoryImplementation<entityType>
     private _dataProvider: DataProvider,
     private _info: EntityFullInfo<entityType>,
     private _defaultFindOptions?: FindOptions<entityType>,
-  ) { }
+  ) {}
   _idCache = new Map<any, any>()
   _getCachedById(id: any, doNotLoadIfNotFound: boolean): entityType {
     id = id + ''
@@ -526,12 +526,12 @@ export class RepositoryImplementation<entityType>
         if (typeof l === 'function') {
           listener = {
             next: l,
-            complete: () => { },
-            error: () => { },
+            complete: () => {},
+            error: () => {},
           }
         }
-        listener.error ??= () => { }
-        listener.complete ??= () => { }
+        listener.error ??= () => {}
+        listener.complete ??= () => {}
         return this._remult.liveQuerySubscriber.subscribe(
           this,
           options,
@@ -744,8 +744,8 @@ export class RepositoryImplementation<entityType>
       let val =
         rel.type === 'reference'
           ? (
-            getEntityRef(row).fields.find(field.key) as IdFieldRef<any, any>
-          ).getId()
+              getEntityRef(row).fields.find(field.key) as IdFieldRef<any, any>
+            ).getId()
           : row[key]
       if (rel.type === 'toOne' || rel.type === 'reference') {
         if (val === null) returnNull = true
@@ -865,7 +865,11 @@ export class RepositoryImplementation<entityType>
   async count(where?: EntityFilter<entityType>): Promise<number> {
     return this._edp.count(await this._translateWhereToFilter(where))
   }
-  async deleteMany({ where }: { where: EntityFilter<entityType> }): Promise<number> {
+  async deleteMany({
+    where,
+  }: {
+    where: EntityFilter<entityType>
+  }): Promise<number> {
     Filter.throwErrorIfFilterIsEmpty(where, 'deleteMany')
     if (this._dataProvider.isProxy) {
       return (this._edp as any as ProxyEntityDataProvider).deleteMany(
@@ -1031,13 +1035,17 @@ export class RepositoryImplementation<entityType>
       if (this.metadata.options.backendPreprocessFilter) {
         where = await this.metadata.options.backendPreprocessFilter(where, {
           metadata: this.metadata,
-          getFilterInfo: (filter) => Filter.getInfo(this.metadata, filter || where)
+          getFilterInfo: (filter) =>
+            Filter.getInfo(this.metadata, filter || where),
         })
       }
       if (this.metadata.options.backendPrefilter) {
         let z = where
         where = {
-          $and: [z, await Filter.resolve(this.metadata.options.backendPrefilter)],
+          $and: [
+            z,
+            await Filter.resolve(this.metadata.options.backendPrefilter),
+          ],
         } as EntityFilter<entityType>
       }
     }
@@ -1347,7 +1355,7 @@ abstract class rowHelperBase<T> {
     this.originalValues = this.copyDataToObject()
     this.saveMoreOriginalData()
   }
-  saveMoreOriginalData() { }
+  saveMoreOriginalData() {}
   async validate() {
     this.__clearErrorsAndReportChanged()
     if (classValidatorValidate)
@@ -1364,7 +1372,7 @@ abstract class rowHelperBase<T> {
     await this.__performColumnAndEntityValidations()
     this.__assertValidity()
   }
-  async __performColumnAndEntityValidations() { }
+  async __performColumnAndEntityValidations() {}
   toApiJson(includeRelatedEntities = false, notJustApi = false) {
     let result: any = {}
     for (const col of this.fieldsMetadata) {
@@ -1671,7 +1679,7 @@ export class rowHelperImplementation<T>
     return d
   }
 
-  private buildLifeCycleEvent(preventDefault: VoidFunction = () => { }) {
+  private buildLifeCycleEvent(preventDefault: VoidFunction = () => {}) {
     const self = this
     return {
       isNew: self.isNew(),
@@ -1787,7 +1795,7 @@ export class rowHelperImplementation<T>
     }
 
     if (this.info.entityInfo.validation) {
-      let e = this.buildLifeCycleEvent(() => { })
+      let e = this.buildLifeCycleEvent(() => {})
       await this.info.entityInfo.validation(this.instance, e)
     }
     if (this.repository.listeners)
@@ -1929,11 +1937,11 @@ export class FieldRefImplementation<entityType, valueType>
       if (rel.type === 'toMany') {
         return (this.container[this.metadata.key] = await this.helper.repository
           .relations(this.container)
-        [this.metadata.key].find())
+          [this.metadata.key].find())
       } else {
         let val = await this.helper.repository
           .relations(this.container)
-        [this.metadata.key].findOne()
+          [this.metadata.key].findOne()
         if (val) this.container[this.metadata.key] = val
         else return null
       }
@@ -2166,8 +2174,9 @@ export class columnDefsImpl implements FieldMetadata {
             } catch (err: any) {
               const error = `${String(
                 prop,
-              )} failed for value ${args?.[0]}. Error: ${typeof err === 'string' ? err : err.message
-                }`
+              )} failed for value ${args?.[0]}. Error: ${
+                typeof err === 'string' ? err : err.message
+              }`
               throw {
                 message: this.caption + ': ' + error,
                 modelState: {
@@ -2358,9 +2367,12 @@ class EntityFullInfo<T> implements EntityMetadata<T> {
     getIdFilter: (...ids: any[]): EntityFilter<any> => {
       if (this.idMetadata.field instanceof CompoundIdField) {
         let field = this.idMetadata.field
-        return {
-          $or: ids.map((x) => field.isEqualTo(x)),
-        }
+        if (ids.length == 1) {
+          return field.isEqualTo(ids[0])
+        } else
+          return {
+            $or: ids.map((x) => field.isEqualTo(x)),
+          }
       }
       if (ids.length == 1)
         return {
@@ -2409,9 +2421,10 @@ export function ValueListFieldType<valueType extends ValueListItem = any>(
   return (type: ClassType<valueType>, context?) => {
     FieldType<valueType>(
       (o) => {
-        ; (o.valueConverter = ValueListInfo.get(type)),
+        ;(o.valueConverter = ValueListInfo.get(type)),
           (o.displayValue = (item, val) => val?.caption)
-        o.validate = (entity, ref) => Validators.in(ValueListInfo.get(type).getValues())(entity, ref)
+        o.validate = (entity, ref) =>
+          Validators.in(ValueListInfo.get(type).getValues())(entity, ref)
       },
       ...options,
     )(type, context)
@@ -2474,7 +2487,7 @@ export class ValueListInfo<T extends ValueListItem>
     } else
       throw new Error(
         `ValueType not yet initialized, did you forget to call @ValueListFieldType on ` +
-        valueListType,
+          valueListType,
       )
   }
 
@@ -2543,7 +2556,7 @@ export function buildOptions<entityType = any, valueType = any>(
     if (o) {
       if (typeof o === 'function') o(r, remult)
       else {
-        const { validate, ...otherOptions } = o;
+        const { validate, ...otherOptions } = o
         r.validate = addValidator(r.validate, validate)
         Object.assign(r, otherOptions)
       }
@@ -2560,7 +2573,6 @@ export function decorateColumnSettings<valueType>(
     let settingsOnTypeLevel = settings.valueType[storableMember]
     if (settingsOnTypeLevel) {
       settings = buildOptions([...settingsOnTypeLevel, settings], remult)
-
     }
   }
 
@@ -2832,9 +2844,9 @@ class SubscribableImp implements Subscribable {
     listener:
       | (() => void)
       | {
-        reportChanged: () => void
-        reportObserved: () => void
-      },
+          reportChanged: () => void
+          reportObserved: () => void
+        },
   ): Unsubscribe {
     let list: {
       reportChanged: () => void
@@ -2843,7 +2855,7 @@ class SubscribableImp implements Subscribable {
     if (typeof listener === 'function')
       list = {
         reportChanged: () => listener(),
-        reportObserved: () => { },
+        reportObserved: () => {},
       }
     else list = listener