+++ projects/tests/dbs/sql-js.spec.ts
@@ -1,10 +1,11 @@
 import { it, describe, expect, beforeEach } from 'vitest'
-import { Remult, SqlDatabase } from '../../core'
+import { Entity, Fields, Remult, SqlDatabase } from '../../core'
 import initSqlJs from 'sql.js'
 import { SqlJsDataProvider } from '../../core/remult-sql-js.js'
 import { allDbTests } from './shared-tests'
 import { SqlDbTests } from './shared-tests/sql-db-tests.js'
 import type { DbTestProps } from './shared-tests/db-tests-props.js'
+import { testMigrationScript } from '../tests/testHelper.js'
 
 describe('Sql JS', () => {
   let db: SqlDatabase
@@ -23,7 +24,7 @@ describe('Sql JS', () => {
       return remult
     },
     createEntity: async (entity) => {
-      const repo = remult.repo(entity);
+      const repo = remult.repo(entity)
       await db.ensureSchema([repo.metadata])
       return repo
     },
@@ -42,13 +43,8 @@ describe('Sql JS', () => {
       ]
     `)
     const c = db.createCommand()
-    expect(
-      (
-        await c.execute(
-          `select * from x where id=${c.param(1)}`,
-        )
-      ).rows,
-    ).toMatchInlineSnapshot(`
+    expect((await c.execute(`select * from x where id=${c.param(1)}`)).rows)
+      .toMatchInlineSnapshot(`
       [
         {
           "id": 1,
@@ -56,5 +52,27 @@ describe('Sql JS', () => {
       ]
     `)
   })
-})
+  it('test knex storage', async () => {
+    @Entity('my')
+    class MyEntity {
+      @Fields.string()
+      name = ''
+      @Fields.json()
+      json = []
+      @Fields.object()
+      object = []
+    }
+    const e = remult.repo(MyEntity).metadata
 
+    expect(
+      await testMigrationScript(db, (m) => m.addColumn(e, e.fields.object)),
+    ).toMatchInlineSnapshot(
+      '"alter table my add column object text default \'\' not null "',
+    )
+    expect(
+      await testMigrationScript(db, (m) => m.addColumn(e, e.fields.name)),
+    ).toMatchInlineSnapshot(
+      '"alter table my add column name text default \'\' not null "',
+    )
+  })
+})